Crop moving image around PTV before registration (#45)
* Add moving image cropping around PTV

* Adjust ROI matching and registration parameters

Changed ROI name matching to be case-insensitive and updated the prefix to lowercase. Modified slice range calculation to expand the range by 20 units on both ends and added debug print statements. Reduced metric sampling percentage and number of iterations in the registration method, and commented out shrink and smoothing settings for flexibility.